refactor: split renderer into gfx/ (camera, types, mesh, pipeline, util, shader.wgsl); thorough comments; compile clean
