+++ src/gfx/pipeline.rs
@@ -1,0 +1,103 @@
+//! Pipeline creation helpers and shader loading.
+//!
+//! WGSL source lives in `shader.wgsl` next to this file and is embedded at compile time
+//! with `include_str!` for convenience (no runtime file IO).
+
+use wgpu::{BindGroupLayout, ColorTargetState, FragmentState, PipelineLayoutDescriptor, PolygonMode, RenderPipeline, ShaderModule, ShaderSource, VertexState};
+
+use crate::gfx::types::{Instance, Vertex};
+
+pub fn create_shader(device: &wgpu::Device) -> ShaderModule {
+    device.create_shader_module(wgpu::ShaderModuleDescriptor {
+        label: Some("basic-shader"),
+        source: ShaderSource::Wgsl(std::borrow::Cow::Borrowed(include_str!("shader.wgsl"))),
+    })
+}
+
+pub fn create_bind_group_layouts(device: &wgpu::Device) -> (BindGroupLayout, BindGroupLayout) {
+    // Globals (view/proj + time)
+    let globals = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+        label: Some("globals-bgl"),
+        entries: &[wgpu::BindGroupLayoutEntry {
+            binding: 0,
+            visibility: wgpu::ShaderStages::VERTEX | wgpu::ShaderStages::FRAGMENT,
+            ty: wgpu::BindingType::Buffer { ty: wgpu::BufferBindingType::Uniform, has_dynamic_offset: false, min_binding_size: None },
+            count: None,
+        }],
+    });
+
+    // Per-draw Model
+    let model = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
+        label: Some("model-bgl"),
+        entries: &[wgpu::BindGroupLayoutEntry {
+            binding: 0,
+            visibility: wgpu::ShaderStages::VERTEX | wgpu::ShaderStages::FRAGMENT,
+            ty: wgpu::BindingType::Buffer { ty: wgpu::BufferBindingType::Uniform, has_dynamic_offset: false, min_binding_size: None },
+            count: None,
+        }],
+    });
+
+    (globals, model)
+}
+
+pub fn create_pipelines(
+    device: &wgpu::Device,
+    shader: &ShaderModule,
+    globals_bgl: &BindGroupLayout,
+    model_bgl: &BindGroupLayout,
+    color_format: wgpu::TextureFormat,
+) -> (RenderPipeline, RenderPipeline, Option<RenderPipeline>) {
+    let pipeline_layout = device.create_pipeline_layout(&PipelineLayoutDescriptor {
+        label: Some("pipeline-layout"),
+        bind_group_layouts: &[globals_bgl, model_bgl],
+        push_constant_ranges: &[],
+    });
+
+    let depth_format = wgpu::TextureFormat::Depth32Float;
+
+    // Non-instanced pipeline (plane)
+    let pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
+        label: Some("pipeline"),
+        layout: Some(&pipeline_layout),
+        vertex: VertexState { module: shader, entry_point: Some("vs_main"), buffers: &[Vertex::LAYOUT], compilation_options: Default::default() },
+        fragment: Some(FragmentState { module: shader, entry_point: Some("fs_main"), targets: &[Some(ColorTargetState { format: color_format, blend: Some(wgpu::BlendState::ALPHA_BLENDING), write_mask: wgpu::ColorWrites::ALL })], compilation_options: Default::default() }),
+        primitive: wgpu::PrimitiveState::default(),
+        depth_stencil: Some(wgpu::DepthStencilState { format: depth_format, depth_write_enabled: true, depth_compare: wgpu::CompareFunction::Less, stencil: wgpu::StencilState::default(), bias: wgpu::DepthBiasState::default() }),
+        multisample: wgpu::MultisampleState::default(),
+        multiview: None,
+        cache: None,
+    });
+
+    // Instanced pipeline (adds per-instance buffer)
+    let inst_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
+        label: Some("inst-pipeline"),
+        layout: Some(&pipeline_layout),
+        vertex: VertexState { module: shader, entry_point: Some("vs_inst"), buffers: &[Vertex::LAYOUT, Instance::LAYOUT], compilation_options: Default::default() },
+        fragment: Some(FragmentState { module: shader, entry_point: Some("fs_inst"), targets: &[Some(ColorTargetState { format: color_format, blend: Some(wgpu::BlendState::ALPHA_BLENDING), write_mask: wgpu::ColorWrites::ALL })], compilation_options: Default::default() }),
+        primitive: wgpu::PrimitiveState::default(),
+        depth_stencil: Some(wgpu::DepthStencilState { format: depth_format, depth_write_enabled: true, depth_compare: wgpu::CompareFunction::Less, stencil: wgpu::StencilState::default(), bias: wgpu::DepthBiasState::default() }),
+        multisample: wgpu::MultisampleState::default(),
+        multiview: None,
+        cache: None,
+    });
+
+    // Optional wireframe
+    let wire_pipeline = if device.features().contains(wgpu::Features::POLYGON_MODE_LINE) {
+        Some(device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
+            label: Some("inst-pipeline-wire"),
+            layout: Some(&pipeline_layout),
+            vertex: VertexState { module: shader, entry_point: Some("vs_inst"), buffers: &[Vertex::LAYOUT, Instance::LAYOUT], compilation_options: Default::default() },
+            fragment: Some(FragmentState { module: shader, entry_point: Some("fs_inst"), targets: &[Some(ColorTargetState { format: color_format, blend: Some(wgpu::BlendState::ALPHA_BLENDING), write_mask: wgpu::ColorWrites::ALL })], compilation_options: Default::default() }),
+            primitive: wgpu::PrimitiveState { polygon_mode: PolygonMode::Line, ..Default::default() },
+            depth_stencil: Some(wgpu::DepthStencilState { format: depth_format, depth_write_enabled: true, depth_compare: wgpu::CompareFunction::Less, stencil: wgpu::StencilState::default(), bias: wgpu::DepthBiasState::default() }),
+            multisample: wgpu::MultisampleState::default(),
+            multiview: None,
+            cache: None,
+        }))
+    } else {
+        None
+    };
+
+    (pipeline, inst_pipeline, wire_pipeline)
+}
+
