+++ src/gfx/types.rs
@@ -1,0 +1,57 @@
+//! Buffer/vertex types shared across pipelines.
+//!
+//! All types here are `#[repr(C)]` and `bytemuck`-safe so they can be uploaded to GPU buffers
+//! without extra copies.
+
+use bytemuck::{Pod, Zeroable};
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Pod, Zeroable)]
+pub struct Globals {
+    pub view_proj: [[f32; 4]; 4],
+    pub time_pad: [f32; 4],
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Pod, Zeroable)]
+pub struct Model {
+    pub model: [[f32; 4]; 4],
+    pub color: [f32; 3],
+    pub emissive: f32,
+    pub _pad: [f32; 4],
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Pod, Zeroable)]
+pub struct Vertex {
+    pub pos: [f32; 3],
+    pub nrm: [f32; 3],
+}
+
+impl Vertex {
+    pub const LAYOUT: wgpu::VertexBufferLayout<'static> = wgpu::VertexBufferLayout {
+        array_stride: std::mem::size_of::<Vertex>() as u64,
+        step_mode: wgpu::VertexStepMode::Vertex,
+        attributes: &wgpu::vertex_attr_array![0 => Float32x3, 1 => Float32x3],
+    };
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, Pod, Zeroable)]
+pub struct Instance {
+    pub model: [[f32; 4]; 4],
+    pub color: [f32; 3],
+    pub selected: f32,
+}
+
+impl Instance {
+    pub const LAYOUT: wgpu::VertexBufferLayout<'static> = wgpu::VertexBufferLayout {
+        array_stride: std::mem::size_of::<Instance>() as u64,
+        step_mode: wgpu::VertexStepMode::Instance,
+        attributes: &wgpu::vertex_attr_array![
+            2 => Float32x4, 3 => Float32x4, 4 => Float32x4, 5 => Float32x4,
+            6 => Float32x3, 7 => Float32
+        ],
+    };
+}
+
